+++ tools/release/one_click_release.ps1
@@ -49,6 +49,7 @@ param(
   [string]$RbGolden = '1',
   [string]$RbCflagsExtra = '-O3 -s',
   [string]$RbOptProfile = 'lto-auto',
+  [ValidateSet('0','1')][string]$RbPreflight = '0',
   # Support multiple sync dirs separated by ';' or ','
   [string]$RbSyncDir
 )
@@ -150,7 +151,7 @@ if ($doBuild) {
     if ($RbSmokeArgs -and $RbSmokeArgs.Trim() -ne '') { $argSmoke = "-a '$RbSmokeArgs'" }
     $argOpt = ''
     if ($RbOptProfile -and $RbOptProfile.Trim() -ne '') { $argOpt = "-O '$RbOptProfile'" }
-    $rbCmd = "tools/remote/remote_build_and_test.sh -H $RbHost -u $RbUser -k '$RbKey' -r $RbSmoke -q '$RbQueries' -s '$primarySyncDir' -P 1 $argSmoke -G $RbGolden -E '$RbCflagsExtra' $argOpt"
+    $rbCmd = "tools/remote/remote_build_and_test.sh -H $RbHost -u $RbUser -k '$RbKey' -r $RbSmoke -q '$RbQueries' -s '$primarySyncDir' -P 1 $argSmoke -G $RbGolden -E '$RbCflagsExtra' $argOpt -K $RbPreflight"
     Invoke-GitBash $rbCmd
 
     # Stage and commit synced statics if changed